refactor: reorganize processor module and update imports; remove unused processor file
