+++ src/ocr/config/ocr_enum.py
