+++ src/ocr/core/ocr_base_engine.py
@@ -9,7 +9,7 @@ import easyocr
 
 from ..config.ocr_config import OCRConfig
 from ..core.image_handler import ImageHandler
-from ..enums.ocr_enum import OutputFormat
+from ..config.ocr_enum import OutputFormat
 from .pdf_base_handler import PDFHandler
 
 logger = logging.getLogger(__name__)
